requirements.txt
aiogram==3.4.1
aiosqlite==0.19.0
python-dotenv==1.0.0
Pillow==10.2.0
pytesseract==0.3.10
PyPDF2==3.0.1
python-docx==1.1.0
httpx==0.26.0
